ThemeButton.vue: refactoring ThemeButton across components
