+++ src/views/layouts/SidebarNavigation.vue
@@ -1,12 +1,7 @@
 <script setup lang="ts">
     import NavbarLink from '@/components/links/NavbarLink.vue';
     import NavbarLowerlink from '@/components/links/NavbarLowerlink.vue';
-
-    const emit = defineEmits(['toggleTheme']);
-
-    const toggleTheme = (): void => {
-        emit('toggleTheme');
-    };
+    import ThemeButton from '@/components/buttons/ThemeButton.vue';
 </script>
 
 <template>
@@ -55,17 +50,11 @@
                 </svg>
             </NavbarLowerlink>
         </li>
-        <li class="md:hidden relitive my-3 cursor-pointer">
-            <div class="flex gap-x-3 justify-start items-center text-sm leading-5 hover:text-primary" @click="toggleTheme">
-                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4 dark:hidden">
-                    <path stroke-linecap="round" stroke-linejoin="round" d="M12 3v2.25m6.364.386l-1.591 1.591M21 12h-2.25m-.386 6.364l-1.591-1.591M12 18.75V21m-4.773-4.227l-1.591 1.591M5.25 12H3m4.227-4.773L5.636 5.636M15.75 12a3.75 3.75 0 11-7.5 0 3.75 3.75 0 017.5 0z" />
-                </svg>
-                <span class="dark:hidden text-sm">Light Theme</span>
-                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4 hidden dark:block">
-                    <path stroke-linecap="round" stroke-linejoin="round" d="M21.752 15.002A9.718 9.718 0 0118 15.75c-5.385 0-9.75-4.365-9.75-9.75 0-1.33.266-2.597.748-3.752A9.753 9.753 0 003 11.25C3 16.635 7.365 21 12.75 21a9.753 9.753 0 009.002-5.998z" />
-                </svg>
+        <li class="md:hidden relitive my-3 hover:text-primary">
+            <ThemeButton class="flex gap-x-3 justify-start items-center text-sm">
                 <span class="hidden dark:block text-sm">Dark Theme</span>
-            </div>
+                <span class="dark:hidden text-sm">Light Theme</span>
+            </ThemeButton>
         </li>
         <li class="lg:hidden relitive my-3 hover:text-primary">
             <NavbarLowerlink title="Documentation" link-to="/documentation">
